ENG-2171: Moved guiFragment parameters handling into GuiFragmentManager
